machinelearning/ai-gateway/appsync-pattern/pyproject.toml
[project]
name = "appsync-pattern"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "asyncio>=3.4.3",
    "aws-cdk-lib>=2.181.0",
    "boto3>=1.37.2",
    "constructs>=10.4.2",
    "gpl>=0.1.4",
    "gql>=3.5.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "requests-aws4auth>=1.3.1",
    "requests-toolbelt>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.9.7",
]
